runtime: make Cmm_Word pointer-width, add riscv64 %quot and BACKEND=riscv64
Cmm_Word (qc--runtime.h) backed every pc_map_entry field - locations,
the return-address location, span values, register/local/span counts -
and the actual register/stack-pointer values the GC's activation-record
walk reads through it. It was hardcoded "unsigned" (32 bits) for every
backend, unconditionally. qc-- itself already emits every one of these
fields at the target's native pointer width (confirmed from a -riscv64
.s dump: inalloc/outalloc/return_addressp/etc. are all ".dword", and
MKOFFSET(0) shows up as bit 63 set, not bit 31), so on a 64-bit target
the C runtime was reading its own pcmap data at the wrong stride,
corrupting the very first activation any real GC walk performed. Retyped
to uintptr_t and widened pcmap.c's Cmm_loctype/Cmm_as_register/
Cmm_as_offset decoding to match (their GOODBITS macro already derives
from sizeof() of their own local variable, so this generalized for free
once that variable's type did) - a no-op on every 32-bit backend,
confirmed by rerunning every other backend's own tests/run-tiger-*.sh
afterwards and matching their recorded baselines exactly.

arch/riscv64/riscv64.ml + riscv64rec.mlb gain %quot (RV64M's "div",
operating on full 64-bit registers, ported from riscv32rec.mlb's own
32-bit "div" rule) - riscv64's instruction selection had no integer
division at all before this.

runtime/Makefile gains BACKEND=riscv64 (a real riscv64-linux-gnu glibc
cross toolchain exists, so - unlike riscv32 - no picolibc workaround is
needed); riscv64cont.s is the riscv64 sibling of every other backend's
own Cmm_unwindcont_pc stub.

Found and fixed bringing up tests/tiger64/ (bits64 tiger test suite)
against -riscv64 - see that directory's README and
docs/claude_notes/notes_64bits.txt for the full story.

+++ runtime/pcmap.h
@@ -7,7 +7,7 @@ typedef struct coded_location { Cmm_Word l; } location;
 typedef enum { DEAD=0, REGISTER=1, OFFSET=2 } loctype; 
 loctype  Cmm_loctype(location l);
 unsigned Cmm_as_register(location l);
-int      Cmm_as_offset  (location l);
+intptr_t Cmm_as_offset  (location l);
 #define isdead(LOC) ((LOC).l == 0)
 /*x: exported type, procedure, and macro declarations */
 typedef struct {
